+++ src/OpenTelemetry.Resources.Azure/ResourceAttributeConstants.cs
@@ -7,6 +7,7 @@ internal sealed class ResourceAttributeConstants
 {
     // AppService resource attributes
     internal const string AzureAppServiceStamp = "azure.app.service.stamp";
+    internal const string AzureResourceGroupName = "azure.resource_group.name";
 
     // Azure VM resource attributes
     internal const string AzureVmScaleSetName = "azure.vm.scaleset.name";
@@ -31,8 +32,8 @@ internal sealed class ResourceAttributeConstants
     internal const string AzureContainerAppJobExecutionNameEnvVar = "CONTAINER_APP_JOB_EXECUTION_NAME";
 
     // Azure resource attributes constant values
-    internal const string AzureAppServicePlatformValue = "azure_app_service";
+    internal const string AzureAppServicePlatformValue = "azure.app_service";
     internal const string AzureCloudProviderValue = "azure";
-    internal const string AzureVmCloudPlatformValue = "azure_vm";
-    internal const string AzureContainerAppsPlatformValue = "azure_container_apps";
+    internal const string AzureVmCloudPlatformValue = "azure.vm";
+    internal const string AzureContainerAppsPlatformValue = "azure.container_apps";
 }