+++ src/OpenTelemetry.SemanticConventions/Attributes/DestinationAttributes.cs
@@ -15,7 +15,7 @@ namespace OpenTelemetry.SemanticConventions;
 public static class DestinationAttributes
 {
     /// <summary>
-    /// Destination address - domain name if available without reverse DNS lookup; otherwise, IP address or Unix domain socket name.
+    /// Destination address - domain name if available without reverse DNS lookup; otherwise, IP address or UNIX domain socket name.
     /// </summary>
     /// <remarks>
     /// When observed from the source side, and when communicating through an intermediary, <c>destination.address</c> SHOULD represent the destination address behind any intermediaries, for example proxies, if it's available.