+++ src/OpenTelemetry.SemanticConventions/Attributes/RpcAttributes.cs
@@ -115,9 +115,7 @@ public static class RpcAttributes
     /// When the method is not recognized, for example, when the server receives
     /// a request for a method that is not predefined on the server, or when
     /// instrumentation is not able to reliably detect if the method is predefined,
-    /// the attribute MUST be set to <c>_OTHER</c>. In such cases, tracing
-    /// instrumentations MUST also set <c>rpc.method_original</c> attribute to
-    /// the original method value.
+    /// the attribute MUST be set to <c>_OTHER</c>.
     /// <p>
     /// If the RPC instrumentation could end up converting valid RPC methods to
     /// <c>_OTHER</c>, then it SHOULD provide a way to configure the list of recognized